SampleInput, SampleOutput: document when an associated IOChannel is     connected or not. Updated some constructors and removed some     copy constructors.  Removed init() method.
SampleConnectionRequester::connected returns a pointer to a 
    possibly cloned SampleConnectionRequester.

nidas::util::Inet4Address:  use getnameinfo, getaddrinfo in place
    of gethostbyname_r,gethostbyaddr_r, which according to man
    page are obsolete.  All of these can trip a valgrind
    uninitialized variable error when a lookup fails.
    Remove unused CACHE_DNS_LOOKUPS stuff - not used with dynamic DHCP/DNS.

nidas::core::Project: moved environment variable stuff to nidas::util::Process

nidas::core::*Socket: expand environment variables in port number
    strings

tests/serial_sensor: use netstat to find an available port 
    number for testing.  This is not bombproof, but the
    run_test_2.sh should work 99.9% of the time.

tests/network/dns_lookup.cc: some tests of DNS lookups.

+++ src/nidas/dynld/DSC_Event.cc
@@ -64,7 +64,7 @@ void DSC_Event::open(int flags) throw(n_u::IOException,
     init();
 
     struct GPIO_MM_event_config cfg;
-    cfg.latencyUsecs = rint(getLatency() * USECS_PER_SEC);
+    cfg.latencyUsecs = (int) rint(getLatency() * USECS_PER_SEC);
     ioctl(GPIO_MM_EVENT_START,&cfg,sizeof(cfg));
 }
